+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 - Implement the photo sending method
 
 - Improvements to the http service
+
+## [0.1.2] - 2024-04-07 
+
+- Add error handling to methods
+- Implements The Audio, Video, Document sending method
+- Improvement in media delivery Service